+++ include/roq/fbs/encode.h
@@ -340,7 +340,9 @@ auto encode(B &builder, const roq::GatewaySettings &value) {
       value.mbp_max_depth,
       value.mbp_allow_price_inversion,
       value.mbp_allow_fractional_tick_size,
-      value.mbp_allow_remove_non_existing);
+      value.mbp_allow_remove_non_existing,
+      value.oms_download_has_state,
+      value.oms_download_has_routing_id);
 }
 
 template <typename B>